Guess the number between 0 and 100 with 9 available guesses.
The program will generate a random integer number using a random function. 
Input is taken from the user as an integer number. 
If the user's entered number is smaller than the program-generated number, the user will get the message please enter a larger number. 
If the user's entered number is larger than the program-generated number, the user will get the message please enter a smaller number.
If the user's entered number is equal to the program-generated number, the user will get the message please enter a smaller number.